# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/public": {
            "put": {
                "description": "Update Public",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Update Public",
                "parameters": [
                    {
                        "description": "UpdatePublic",
                        "name": "public",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.GetPublicResponse"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Update Public",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to Update public",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Endpoint for creating a new Public",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Create Public",
                "parameters": [
                    {
                        "description": "Public create",
                        "name": "public",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.CreatePublicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created public",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create public",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/public/:id": {
            "get": {
                "description": "Get Public",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Get Public",
                "parameters": [
                    {
                        "description": "Get Public",
                        "name": "public",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.ByIdPublicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Get Public",
                        "schema": {
                            "$ref": "#/definitions/genproto.GetPublicResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to Get public",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete Public",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Delete Public",
                "parameters": [
                    {
                        "description": "DeletePublic",
                        "name": "public",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.ByIdPublicRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Delete Public",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to Delete public",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/publics": {
            "get": {
                "description": "Get All Public",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Public"
                ],
                "summary": "Get All Public",
                "parameters": [
                    {
                        "type": "integer",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully Get All Public",
                        "schema": {
                            "$ref": "#/definitions/genproto.GetAllPublicResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to Get All public",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/publicvote": {
            "put": {
                "description": "Endpoint for updating a Public Vote",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PublicVote"
                ],
                "summary": "Update Public Vote",
                "parameters": [
                    {
                        "description": "Update Public Vote",
                        "name": "publicvote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.UpdatePublicVoteRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated Public Vote",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to update Public Vote",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Endpoint for creating a new Public Vote",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PublicVote"
                ],
                "summary": "Create Public Vote",
                "parameters": [
                    {
                        "description": "Public Vote create",
                        "name": "publicvote",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/genproto.CreatePublicVoteReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created Public Vote",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to create Public Vote",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/publicvote/{id}": {
            "get": {
                "description": "Endpoint for getting a Public Vote by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PublicVote"
                ],
                "summary": "Get Public Vote by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public Vote ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully fetched Public Vote",
                        "schema": {
                            "$ref": "#/definitions/genproto.PublicVote"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to get Public Vote",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Endpoint for deleting a Public Vote",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PublicVote"
                ],
                "summary": "Delete Public Vote",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Public Vote ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully deleted Public Vote",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to delete Public Vote",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/publicvotes": {
            "get": {
                "description": "Endpoint for getting all Public Votes",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "PublicVote"
                ],
                "summary": "Get All Public Votes",
                "responses": {
                    "200": {
                        "description": "Successfully fetched Public Votes",
                        "schema": {
                            "$ref": "#/definitions/genproto.GetPublicVoteResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Failed to get Public Votes",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "genproto.ByIdPublicRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
...
        "genproto.CreatePublicRequest": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "nation": {
                    "type": "string"
                },
                "party_id": {
                    "type": "string"
                }
            }
        },
        "genproto.CreatePublicVoteReq": {
            "type": "object",
            "properties": {
                "candidate_id": {
                    "type": "string"
                },
                "election_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                }
            }
        },
        "genproto.Election": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
...
        "genproto.GetAllPublicResponse": {
            "type": "object",
            "properties": {
                "publics": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.GetPublicResponse"
                    }
                }
            }
        },
...
        "genproto.GetPublicResponse": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "gender": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "nation": {
                    "type": "string"
                },
                "party_id": {
                    "type": "string"
                }
            }
        },
        "genproto.GetPublicVoteResponse": {
            "type": "object",
            "properties": {
                "public_vote": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/genproto.PublicVote"
                    }
                }
            }
        },
        "genproto.PublicVote": {
            "type": "object",
            "properties": {
                "election": {
                    "$ref": "#/definitions/genproto.Election"
                },
                "id": {
                    "type": "string"
                },
                "public_id": {
                    "type": "string"
                }
            }
        },
...
        "genproto.UpdatePublicVoteRequest": {
            "type": "object",
            "properties": {
                "public_vote": {
                    "$ref": "#/definitions/genproto.CreatePublicVoteReq"
                }
            }
        }
    }
}